Move mocks into individual files to be more manageable
